requirements.txt
web3==5.31.3
face-recognition==1.3.0
pycryptodome==3.15.0
merkletools==1.0.3
py_ecc==5.2.0
opencv-python==4.6.0.66